Add missing EIGEN_DEPRECATED annotations to deprecated functions and fix few other doxygen warnings

+++ unsupported/Eigen/CXX11/src/Tensor/TensorCostModel.h
@@ -177,7 +177,8 @@ class TensorCostModel {
     double threads = (cost - kStartupCycles) / kPerThreadCycles + 0.9;
     // Make sure we don't invoke undefined behavior when we convert to an int.
     threads = numext::mini<double>(threads, GenericNumTraits<int>::highest());
-    return numext::mini(max_threads, numext::maxi<int>(1, threads));
+    return numext::mini(max_threads,
+                        numext::maxi<int>(1, static_cast<int>(threads)));
   }
 
   // taskSize assesses parallel task size.
